chore(deps): rename useStore hook
