+++ packages/big-design/src/components/Worksheet/hooks/index.ts
@@ -2,6 +2,6 @@ export * from './useEditableCell';
 export * from './useExpandable';
 export * from './useKeyEvents';
 export * from './useNavigation';
-export * from './useStore';
+export * from './useWorksheetStore';
 export * from './useTableFocus';
 export * from './useUpdateItems';
