--- packages/big-design/src/components/Worksheet/hooks/useStore/index.ts
@@ -1,1 +1,0 @@
-export * from './useStore';
